Fix Travis

+++ lib/src/marshall.c
@@ -68,7 +68,8 @@ uint8_t varint_size(const uint64_t val)
 
 void enc1(uint8_t ** pos, const uint8_t * const end, const uint8_t val)
 {
-    ensure(*pos + sizeof(val) <= end, "buffer overflow: %zu", end - *pos);
+    ensure(*pos + sizeof(val) <= end, "buffer overflow: %lu",
+           (unsigned long)(end - *pos));
     **pos = val;
     *pos += sizeof(val);
 }
@@ -76,7 +77,8 @@ void enc1(uint8_t ** pos, const uint8_t * const end, const uint8_t val)
 
 void enc2(uint8_t ** pos, const uint8_t * const end, const uint16_t val)
 {
-    ensure(*pos + sizeof(val) <= end, "buffer overflow: %zu", end - *pos);
+    ensure(*pos + sizeof(val) <= end, "buffer overflow: %lu",
+           (unsigned long)(end - *pos));
     const uint16_t v = bswap16(val);
     memcpy(*pos, &v, sizeof(v));
     *pos += sizeof(val);
@@ -85,7 +87,8 @@ void enc2(uint8_t ** pos, const uint8_t * const end, const uint16_t val)
 
 void enc3(uint8_t ** pos, const uint8_t * const end, const uint32_t val)
 {
-    ensure(*pos + 3 <= end, "buffer overflow: %zu", end - *pos);
+    ensure(*pos + 3 <= end, "buffer overflow: %lu",
+           (unsigned long)(end - *pos));
     const uint32_t v = bswap32(val << 8);
     memcpy(*pos, &v, 3);
     *pos += 3;
@@ -94,7 +97,8 @@ void enc3(uint8_t ** pos, const uint8_t * const end, const uint32_t val)
 
 void enc4(uint8_t ** pos, const uint8_t * const end, const uint32_t val)
 {
-    ensure(*pos + sizeof(val) <= end, "buffer overflow: %zu", end - *pos);
+    ensure(*pos + sizeof(val) <= end, "buffer overflow: %lu",
+           (unsigned long)(end - *pos));
     const uint32_t v = bswap32(val);
     memcpy(*pos, &v, sizeof(v));
     *pos += sizeof(val);
@@ -103,7 +107,8 @@ void enc4(uint8_t ** pos, const uint8_t * const end, const uint32_t val)
 
 void enc8(uint8_t ** pos, const uint8_t * const end, const uint64_t val)
 {
-    ensure(*pos + sizeof(val) <= end, "buffer overflow: %zu", end - *pos);
+    ensure(*pos + sizeof(val) <= end, "buffer overflow: %lu",
+           (unsigned long)(end - *pos));
     const uint64_t v = bswap64(val);
     memcpy(*pos, &v, sizeof(v));
     *pos += sizeof(val);
@@ -115,7 +120,8 @@ void encv(uint8_t ** pos, const uint8_t * const end, const uint64_t val)
     ensure((val & VARINT_MASK) == 0, "value overflow: %" PRIu64, val);
 
     if ((val & VARINT_MASK8) != 0) {
-        ensure(*pos + 8 <= end, "buffer overflow: %zu", end - *pos);
+        ensure(*pos + 8 <= end, "buffer overflow: %lu",
+               (unsigned long)(end - *pos));
         *(*pos + 0) = ((val >> 56) & 0x3f) + 0xc0;
         *(*pos + 1) = (val >> 48) & 0xff;
         *(*pos + 2) = (val >> 40) & 0xff;
@@ -129,7 +135,8 @@ void encv(uint8_t ** pos, const uint8_t * const end, const uint64_t val)
     }
 
     if ((val & VARINT_MASK4) != 0) {
-        ensure(*pos + 4 <= end, "buffer overflow: %zu", end - *pos);
+        ensure(*pos + 4 <= end, "buffer overflow: %lu",
+               (unsigned long)(end - *pos));
         *(*pos + 0) = ((val >> 24) & 0x3f) + 0x80;
         *(*pos + 1) = (val >> 16) & 0xff;
         *(*pos + 2) = (val >> 8) & 0xff;
@@ -139,14 +146,16 @@ void encv(uint8_t ** pos, const uint8_t * const end, const uint64_t val)
     }
 
     if ((val & VARINT_MASK2) != 0) {
-        ensure(*pos + 2 <= end, "buffer overflow: %zu", end - *pos);
+        ensure(*pos + 2 <= end, "buffer overflow: %lu",
+               (unsigned long)(end - *pos));
         *(*pos + 0) = ((val >> 8) & 0x3f) + 0x40;
         *(*pos + 1) = val & 0xff;
         *pos += 2;
         return;
     }
 
-    ensure(*pos + 1 <= end, "buffer overflow: %zu", end - *pos);
+    ensure(*pos + 1 <= end, "buffer overflow: %lu",
+           (unsigned long)(end - *pos));
     **pos = val & 0x3f;
     *pos += 1;
 }
@@ -193,7 +202,8 @@ void encb(uint8_t ** pos,
           const uint8_t * const val,
           const uint16_t len)
 {
-    ensure(*pos + len <= end, "buffer overflow: %zu", end - *pos);
+    ensure(*pos + len <= end, "buffer overflow: %lu",
+           (unsigned long)(end - *pos));
     memcpy(*pos, val, len);
     *pos += len;
 }
